Finish day 19
Very interestingly, the initial version WAS slow.  But I was being a
bit slow too, because once I had implemented this faster version of
reducing the molecule instead of building it, I then asked it to
generate ALL solutions instead of just getting the first.  Since I
generate the sequence from shortest to longest, the first will always
be the shortest!